backend/requirements.txt
# Web framework
fastapi>=0.110
uvicorn[standard]>=0.27

# LLM / Embedding client
httpx>=0.27
openai>=1.30

# Vector store
chromadb>=0.5

# Document processing
langchain-core>=0.3
langchain-text-splitters>=0.3
pypdf>=4.0
python-docx>=1.1

# Database
aiosqlite>=0.20

# Config / Logging / Rate limit
pydantic-settings>=2.2
loguru>=0.7
slowapi>=0.1
pyyaml>=6.0

# Testing
pytest>=8.2
pytest-asyncio>=0.23
